+++ app-backend/BackEndTests/Providers/TestableDateTimeProvider.cs
@@ -1,9 +1,9 @@
 ﻿using backend.Controllers;
 
-namespace BackEndTests;
+namespace BackEndTests.Providers;
 
 public class TestableDateTimeProvider : DateTimeProvider
 {
-    protected override DateTime GetUtcNow() => 
+    protected override DateTime GetUtcNow() =>
         DateTime.Parse("2021-05-12T18:29:30.206924+00:00");
 }
