+++ app-backend/backend.sln.DotSettings.user
@@ -1,10 +1,7 @@
 ﻿<wpf:ResourceDictionary xml:space="preserve" xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml" xmlns:s="clr-namespace:System;assembly=mscorlib" xmlns:ss="urn:shemas-jetbrains-com:settings-storage-xaml" xmlns:wpf="http://schemas.microsoft.com/winfx/2006/xaml/presentation">
-	<s:String x:Key="/Default/Environment/UnitTesting/UnitTestSessionStore/Sessions/=625c617d_002D7b14_002D43be_002Db121_002D797e21433561/@EntryIndexedValue">&lt;SessionState ContinuousTestingMode="0" IsActive="True" Name="PingControllerShould" xmlns="urn:schemas-jetbrains-com:jetbrains-ut-session"&gt;&#xD;
+	<s:String x:Key="/Default/Environment/UnitTesting/UnitTestSessionStore/Sessions/=477dfd89_002D16ea_002D4967_002D960a_002D55c7ae396fa9/@EntryIndexedValue">&lt;SessionState ContinuousTestingMode="0" IsActive="True" Name="AccountsControllerShould" xmlns="urn:schemas-jetbrains-com:jetbrains-ut-session"&gt;&#xD;
   &lt;TestAncestor&gt;&#xD;
-    &lt;TestId&gt;NUnit3x::352EF17C-BB4D-405A-875B-B6AC2E58C470::net7.0::BackEndTests.PingControllerShould&lt;/TestId&gt;&#xD;
-    &lt;TestId&gt;NUnit3x::352EF17C-BB4D-405A-875B-B6AC2E58C470::net7.0::BackEndTests.TransactionsControllerShould.ReturnCreatedStatusAndTransactionTypeWhenSaveTransaction&lt;/TestId&gt;&#xD;
-    &lt;TestId&gt;NUnit3x::352EF17C-BB4D-405A-875B-B6AC2E58C470::net7.0::BackEndTests.TransactionsControllerShould&lt;/TestId&gt;&#xD;
-    &lt;TestId&gt;NUnit3x::352EF17C-BB4D-405A-875B-B6AC2E58C470::net7.0::BackEndTests.TransactionRepositoryShould&lt;/TestId&gt;&#xD;
     &lt;TestId&gt;NUnit3x::352EF17C-BB4D-405A-875B-B6AC2E58C470::net7.0::BackEndTests.AccountsControllerShould&lt;/TestId&gt;&#xD;
+    &lt;TestId&gt;NUnit3x::352EF17C-BB4D-405A-875B-B6AC2E58C470::net7.0::BackEndTests.PingControllerShould&lt;/TestId&gt;&#xD;
   &lt;/TestAncestor&gt;&#xD;
 &lt;/SessionState&gt;</s:String></wpf:ResourceDictionary>
